// apps/web/lib/api/types/TradeSimulationRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { TradeLeg } from "./TradeLeg"

/**
 * TradeSimulationRequest
 */
export type TradeSimulationRequest = {
  /**
   * @type array | undefined
   */
  give?: TradeLeg[]
  /**
   * @type array | undefined
   */
  receive?: TradeLeg[]
}